[bugfix] Fixed the bug in memory pool register.
The MemAllocType enum relys on the NumMemTypes to act as the size of the
memory pools.
Adding ExtraVHandlePool at the end of the enum results in the size of
g_ps to be too small and causes seg fault when registering.
Fixed by putting NumMemTypes at the end and added comment to remind.